+++ parka/lib/components/drawer/private-drawer/private_drawer_n.dart
@@ -5,8 +5,8 @@ import 'package:ParkA/pages/edit-profile/edit_profile_page.dart';
 import 'package:ParkA/pages/parkings/parking_page.dart';
 import 'package:ParkA/pages/profile/components/profile_personal_information_widget.dart';
 import 'package:ParkA/pages/login/login_screen.dart';
-import 'package:ParkA/pages/reservation-as-owner/reservation_as_owner.dart';
-import 'package:ParkA/pages/reservations-as-client/reservation_as_client.dart';
+import 'package:ParkA/pages/reservation/reservation-as-owner/reservation_as_owner.dart';
+import 'package:ParkA/pages/reservation/reservations-as-client/reservation_as_client.dart';
 import 'package:ParkA/pages/search/search_panel.dart';
 import 'package:ParkA/pages/user-payments/user_registered_payment_methods_screen.dart';
 import 'package:ParkA/pages/vehicles/vehicle_page.dart';
