+++ parka/lib/pages/reservation/reservations-as-client/reservation_as_client.dart
@@ -4,6 +4,7 @@ import 'package:ParkA/data/data-models/reservation/reservation_data_model.dart';
 import 'package:ParkA/data/enums/parking_place_holder_type.dart';
 import 'package:ParkA/data/use-cases/reservation/reservation_use_cases.dart';
 import 'package:ParkA/pages/profile/components/parka_circle_avatar_widget.dart';
+import 'package:ParkA/pages/reservation/components/profile_widget_tab.dart';
 import 'package:ParkA/styles/parka_colors.dart';
 import 'package:ParkA/styles/text.dart';
 import 'package:auto_size_text/auto_size_text.dart';
@@ -365,107 +366,6 @@ class ActionButtonsOwnerState extends StatelessWidget {
                 ),
               ),
       ),
-    );
-  }
-}
-
-class ProfileWidgetTab extends StatelessWidget {
-  const ProfileWidgetTab({
-    Key key,
-    @required Parking parking,
-  })  : _parking = parking,
-        super(key: key);
-
-  final Parking _parking;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Padding(
-          padding: const EdgeInsets.symmetric(vertical: 8.0),
-          child: Text(
-            "Cliente",
-            style: kParkaTextStyleBoldGreen18,
-          ),
-        ),
-        InkWell(
-          onTap: () {},
-          child: Row(
-            children: [
-              Expanded(
-                flex: 1,
-                child: Center(
-                  child: ParkaCircleAvatarWidget(
-                    imageUrl:
-                        "https://parka-api-bucket-aws.s3.amazonaws.com/pp_857565fdc3.jfif",
-                  ),
-                ),
-              ),
-              Expanded(
-                flex: 6,
-                child: Padding(
-                  padding: const EdgeInsets.only(left: 8.0),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: [
-                      Container(
-                        alignment: Alignment.bottomLeft,
-                        child: Text(
-                          'David Bujosa',
-                          style: kParkaTextStyleBlack18,
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.only(top: 8.0),
-                        child: Row(
-                          children: [
-                            Text(
-                              '4.20 ',
-                              style: kParkaTextStyleBlack18,
-                            ),
-                            Icon(
-                              Icons.star,
-                              color: Colors.black,
-                            ),
-                          ],
-                        ),
-                      )
-                    ],
-                  ),
-                ),
-              )
-            ],
-          ),
-        ),
-        true != true
-            ? Container()
-            : InkWell(
-                onTap: () {},
-                child: Padding(
-                  padding: const EdgeInsets.only(top: 12.0, bottom: 8.0),
-                  child: Container(
-                    child: Row(
-                      children: [
-                        Text(
-                          "Contatar ",
-                          style: TextStyle(
-                              fontFamily: "Montserrat",
-                              color: Color(0xff077187),
-                              fontSize: 17.0),
-                        ),
-                        Icon(
-                          Icons.message,
-                          color: Color(0xff077187),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-      ],
     );
   }
 }
